Add windowed mode (#83)
## The Problem

A number of users on various platforms have run into issues with the
transparent window that displays the widget. Many of these issues have
completely prevented these users from using Swivvel. It would be nice if
we had any workaround for these users, no matter how blunt.

## The Solution

Add a new "Windowed Mode" tray menu option that runs the transparent
window as a regular, non-transparent window. This isn't the ideal way to
run the app, but it's better than not being able to run the app at all.