Add deep link handler to join audio room for pod (#84)
## The Problem

We are building a feature that posts a Slack message when a new Swivvel
session is started. This message will have a link to join the Swivvel
session. This requires a deep link in to the Electron app that runs the
code to join the audio room for a specific pod.

## The Solution

Add a new `/join-audio-room-for-pod?podId=xxx` deep link that sends the
existing `joinAudioRoomForPod` event to the transparent window.

This commit also fixes a bug where we were slicing the last character
off of the deep link when we should not have been.